Updated account transfer money implementation & Implemented transaction history

+++ src/main/java/com/banquemisr/moneytransactionservice/repository/TransactionRepository.java
@@ -4,6 +4,7 @@ import com.banquemisr.moneytransactionservice.model.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 }
